Homework 6.3 Improved

+++ 06_InheritanceAndPolymorphism/AbstractClient/src/Main.java
@@ -2,21 +2,22 @@ import clients.BusinessPerson;
 import clients.Company;
 import clients.Person;
 
+import java.math.BigDecimal;
 import java.math.BigInteger;
 import java.util.concurrent.ThreadLocalRandom;
 
 public class Main {
     public static void main(String[] args) {
         Person pupkin = new Person("Вася Пупкин");
-        pupkin.deposit(10000.0);
-        pupkin.withdraw(5000.0);
+        pupkin.deposit(BigDecimal.valueOf(10000.0));
+        pupkin.withdraw(BigDecimal.valueOf(5000.0));
         System.out.println(pupkin.getBalance());
         System.out.println(pupkin);
 
         BusinessPerson dengoff = new BusinessPerson("Иван Деньгов");
-        dengoff.deposit(10000.0);
-        dengoff.deposit(500.0);
-        dengoff.withdraw(5000.0);
+        dengoff.deposit(BigDecimal.valueOf(10000.0));
+        dengoff.deposit(BigDecimal.valueOf(500.0));
+        dengoff.withdraw(BigDecimal.valueOf(5000.0));
         System.out.println(dengoff.getBalance());
         System.out.println(dengoff);
 
@@ -26,8 +27,8 @@ public class Main {
         }
         BigInteger inn = new BigInteger(innNumber.toString());
         Company mmm = new Company("МММ", inn);
-        mmm.deposit(10000.0);
-        mmm.withdraw(5000.0);
+        mmm.deposit(BigDecimal.valueOf(10000.0));
+        mmm.withdraw(BigDecimal.valueOf(5000.0));
         System.out.println(mmm.getBalance());
         System.out.println(mmm);
     }
